pyproject.toml
[tool.poetry]
name = "referral-system"
version = "0.1.0"
description = ""
authors = ["Azamat Baltaev <1210044@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-environ = "^0.11.2"
drf-yasg = "^1.21.8"
phonenumbers = "^8.13.50"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
